+++ dogen-ap/CHANGELOG.md
@@ -7,3 +7,7 @@
 ## Version 0.0.7
 - Reduced required parameters by calculating them from the environment.
 - Added support for extension instances.
+
+## Version 0.0.8
+- Make all fields optional when a generation is created.
+- Update updatedAt when a generation is updated.